elementwise_util: don't cast the result of compute_fun back to the common type
The compute function might return an entirely different type. For
example, if we were applying a trigonometric function like acos to an
input of type bool expecting an output of type float, we would get bad
results because acos(0) = 1.57, but casting through bool would
truncate that to 1.

Note that we don't need the pair of ET_CHECK_MSG I removed because we
already check tensor dtypes on entry to the elementwise util
functions; the checks were inconvenient because we now call
get_store_common_to_tensor_fn without the actual common type.

ghstack-source-id: a53266913aa795b57ab4804b2709d9850a296194
ghstack-comment-id: 2735017325
Pull Request resolved: https://github.com/pytorch/executorch/pull/9385